Unify the two truncate helpers and measure display width
truncate_cols (trigger param table) and truncate_breadcrumb_part
(breadcrumb) did the same job with different ellipsis characters, so the
UI clipped text two different ways. Both also counted chars rather than
terminal columns, which under-truncates wide characters: a CJK dag_id or
task_id clipped to N chars renders up to 2N columns and overflows its
cell.

Replace both with a single ui::common::truncate_cols that measures
display width via unicode-width, already a direct dependency used the
same way in ui::tabs.

Breadcrumbs now clip with '…' instead of '...'; the total width budget is
unchanged.
