+++ src/app/state/breadcrumb.rs
@@ -3,6 +3,7 @@ use std::sync::LazyLock;
 use time::format_description::BorrowedFormatItem;
 
 use super::{App, Panel};
+use crate::ui::common::truncate_cols;
 
 /// Cached date format for breadcrumb display (YYYY-MM-DD)
 static BREADCRUMB_DATE_FORMAT: LazyLock<Vec<BorrowedFormatItem<'static>>> = LazyLock::new(|| {
@@ -20,7 +21,7 @@ impl App {
 
         if self.active_panel != Panel::Config && self.active_panel != Panel::Dag {
             if let Some(dag_id) = self.nav_context.dag_id() {
-                let truncated = Self::truncate_breadcrumb_part(dag_id, 25);
+                let truncated = truncate_cols(dag_id, 25);
                 parts.push(truncated);
             }
         }
@@ -33,18 +34,18 @@ impl App {
                         .unwrap_or_else(|_| "unknown".to_string());
                     parts.push(formatted);
                 } else if let Some(dag_run_id) = self.nav_context.dag_run_id() {
-                    let truncated = Self::truncate_breadcrumb_part(dag_run_id, 20);
+                    let truncated = truncate_cols(dag_run_id, 20);
                     parts.push(truncated);
                 }
             } else if let Some(dag_run_id) = self.nav_context.dag_run_id() {
-                let truncated = Self::truncate_breadcrumb_part(dag_run_id, 20);
+                let truncated = truncate_cols(dag_run_id, 20);
                 parts.push(truncated);
             }
         }
 
         if self.active_panel == Panel::Logs {
             if let Some(task_id) = self.nav_context.task_id() {
-                let truncated = Self::truncate_breadcrumb_part(task_id, 20);
+                let truncated = truncate_cols(task_id, 20);
                 parts.push(truncated);
             }
         }
@@ -55,20 +56,6 @@ impl App {
             Some(parts[0].clone())
         } else {
             None
-        }
-    }
-
-    fn truncate_breadcrumb_part(s: &str, max_chars: usize) -> String {
-        let char_count = s.chars().count();
-        if char_count <= max_chars {
-            s.to_string()
-        } else {
-            let truncate_at = max_chars.saturating_sub(3);
-            let byte_index = s
-                .char_indices()
-                .nth(truncate_at)
-                .map_or(s.len(), |(idx, _)| idx);
-            format!("{}...", &s[..byte_index])
         }
     }
 }
